Adds in two human strategies, add in CLI args for configuring the num of humans/goals, and also starts to modify the grid env to account for multiple humans/goals
