+++ humans/human_types.py
@@ -1,0 +1,5 @@
+from enum import IntEnum
+
+class HumanTypes(IntEnum):
+    NOISY_GREEDY = 0
+    RANDOM = 1
